orders_app , reviews_app tests

+++ offers_app/models.py
@@ -42,10 +42,7 @@ class OfferDetail(models.Model):
     title = models.CharField(max_length=255)
     revisions = models.IntegerField()
     delivery_time_in_days = models.IntegerField()
-    price = models.DecimalField(
-        max_digits=10,
-        decimal_places=2,
-    )
+    price = models.IntegerField()
     features = models.JSONField(default=list, blank=True)
     offer_type = models.CharField(max_length=20, choices=OFFER_TYPES)
